+++ src/main/java/gregtech/loaders/postload/GT_MachineRecipeLoader.java
@@ -55,31 +55,7 @@ public class GT_MachineRecipeLoader implements Runnable {
     @Override
     public void run() {
         GT_Log.out.println("GT_Mod: Adding non-OreDict Machine Recipes.");
-        try {
-            GT_Utility.removeSimpleIC2MachineRecipe(
-                    GT_Values.NI,
-                    ic2.api.recipe.Recipes.metalformerExtruding.getRecipes(),
-                    ItemList.Cell_Empty.get(3L));
-            GT_Utility.removeSimpleIC2MachineRecipe(
-                    ItemList.IC2_Energium_Dust.get(1L), ic2.api.recipe.Recipes.compressor.getRecipes(), GT_Values.NI);
-            GT_Utility.removeSimpleIC2MachineRecipe(
-                    new ItemStack(Items.gunpowder), ic2.api.recipe.Recipes.extractor.getRecipes(), GT_Values.NI);
-            GT_Utility.removeSimpleIC2MachineRecipe(
-                    new ItemStack(Blocks.wool, 1, 32767), ic2.api.recipe.Recipes.extractor.getRecipes(), GT_Values.NI);
-            GT_Utility.removeSimpleIC2MachineRecipe(
-                    new ItemStack(Blocks.gravel), ic2.api.recipe.Recipes.oreWashing.getRecipes(), GT_Values.NI);
-        } catch (Throwable ignored) {
-        }
-        GT_Utility.removeIC2BottleRecipe(
-                GT_ModHandler.getIC2Item("fuelRod", 1),
-                GT_ModHandler.getIC2Item("UranFuel", 1),
-                ic2.api.recipe.Recipes.cannerBottle.getRecipes(),
-                GT_ModHandler.getIC2Item("reactorUraniumSimple", 1, 1));
-        GT_Utility.removeIC2BottleRecipe(
-                GT_ModHandler.getIC2Item("fuelRod", 1),
-                GT_ModHandler.getIC2Item("MOXFuel", 1),
-                ic2.api.recipe.Recipes.cannerBottle.getRecipes(),
-                GT_ModHandler.getIC2Item("reactorMOXSimple", 1, 1));
+
 
         try {
             GT_DummyWorld tWorld = (GT_DummyWorld) GT_Values.DW;
@@ -262,7 +238,7 @@ public class GT_MachineRecipeLoader implements Runnable {
                 500,
                 48);
 
-        GT_ModHandler.removeRecipeByOutput(ItemList.IC2_Fertilizer.get(1L));
+
 
         for (int g = 0; g < 16; g++) {
             if (!GT_MachineRecipeLoader.isNEILoaded) {
@@ -285,41 +261,8 @@ public class GT_MachineRecipeLoader implements Runnable {
 
     public void run2() {
 
-        GT_ModHandler.removeRecipe(new ItemStack(Items.lava_bucket), ItemList.Cell_Empty.get(1L));
-        GT_ModHandler.removeRecipe(new ItemStack(Items.water_bucket), ItemList.Cell_Empty.get(1L));
-
-        GT_ModHandler.removeFurnaceSmelting(ItemList.IC2_Resin.get(1L));
-
         GT_NaniteChain.run();
         GT_PCBFactoryRecipes.load();
-
-        GT_Utility.removeSimpleIC2MachineRecipe(
-                new ItemStack(Blocks.cobblestone),
-                GT_ModHandler.getMaceratorRecipeList(),
-                GT_OreDictUnificator.get(OrePrefixes.dust, Materials.Stone, 1L));
-        GT_Utility.removeSimpleIC2MachineRecipe(
-                GT_OreDictUnificator.get(OrePrefixes.gem, Materials.Lapis, 1L),
-                GT_ModHandler.getMaceratorRecipeList(),
-                ItemList.IC2_Plantball.get(1L));
-        GT_Utility.removeSimpleIC2MachineRecipe(
-                GT_OreDictUnificator.get(OrePrefixes.dust, Materials.Redstone, 1L),
-                GT_ModHandler.getMaceratorRecipeList(),
-                ItemList.IC2_Plantball.get(1L));
-        GT_Utility.removeSimpleIC2MachineRecipe(
-                GT_OreDictUnificator.get(OrePrefixes.dust, Materials.Glowstone, 1L),
-                GT_ModHandler.getMaceratorRecipeList(),
-                ItemList.IC2_Plantball.get(1L));
-        GT_Utility.removeSimpleIC2MachineRecipe(
-                GT_Values.NI, GT_ModHandler.getMaceratorRecipeList(), getModItem("IC2", "itemBiochaff", 1L));
-
-        GT_Utility.removeSimpleIC2MachineRecipe(
-                new ItemStack(Blocks.cactus, 8, 0),
-                GT_ModHandler.getCompressorRecipeList(),
-                getModItem("IC2", "itemFuelPlantBall", 1L));
-        GT_Utility.removeSimpleIC2MachineRecipe(
-                getModItem("ExtraTrees", "food", 8L, 24),
-                GT_ModHandler.getCompressorRecipeList(),
-                getModItem("IC2", "itemFuelPlantBall", 1L));
 
         if (!GregTech_API.mIC2Classic) {
             try {
@@ -354,12 +297,6 @@ public class GT_MachineRecipeLoader implements Runnable {
                 /*Do nothing*/
             }
         }
-        GT_Utility.removeSimpleIC2MachineRecipe(
-                ItemList.Crop_Drop_BobsYerUncleRanks.get(1L), GT_ModHandler.getExtractorRecipeList(), null);
-        GT_Utility.removeSimpleIC2MachineRecipe(
-                ItemList.Crop_Drop_Ferru.get(1L), GT_ModHandler.getExtractorRecipeList(), null);
-        GT_Utility.removeSimpleIC2MachineRecipe(
-                ItemList.Crop_Drop_Aurelia.get(1L), GT_ModHandler.getExtractorRecipeList(), null);
 
         ItemStack tCrop;
         // Metals Line
